+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ----------
 
 v1.0.17 (??.??.2016)
+ * dirty state after setProperty([]) should be dirty, not clean
  * new initial abstract detail handler for be modules added
  * new abstract seach handler for be modules added 
  * new abstract lister for be modules added